refactor(tempo): split monolithic module into submodules
Split the 1302-line tempo.rs into:
- mod.rs: Tempo struct, Message/Action enums, update(), view(), subscription()
- calendar.rs: time_str(), naive_date(), calendar(), calendar_with_theme()
- weather.rs: weather() view, fetch_location(), fetch_weather_data(), types, weather_icon()
